+++ src/entities/rss/lib/index.ts
@@ -1,0 +1,15 @@
+import { type RemotePattern, RSS_FEEDS } from '../model';
+
+export const remotePatterns: RemotePattern[] = RSS_FEEDS.map(({ url }) => {
+  const { protocol: fullProtocol, hostname } = new URL(url);
+  const protocol = fullProtocol.replace(':', '');
+
+  return {
+    ...(_isProtocol(protocol) && { protocol }),
+    hostname
+  };
+});
+
+function _isProtocol(protocol: string): protocol is 'http' | 'https' {
+  return ['http', 'https'].includes(protocol.replace(':', ''));
+}
