+++ src/entities/rss/model/index.ts
@@ -1,0 +1,2 @@
+export * from './constants';
+export * from './types';
