+++ src/entities/rss/model/types.ts
@@ -1,0 +1,7 @@
+export type RemotePattern = {
+  protocol?: 'http' | 'https';
+  hostname: string;
+  port?: string;
+  pathname?: string;
+  search?: string;
+};
